pyproject.toml
[tool.poetry]
name = "pythonproject"
version = "0.1.0"
description = ""
authors = ["kirill <w_shkolenko@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.103.0"}
alembic = "^1.11.3"
loguru = "^0.7.0"
ruff = "^0.3.7"
pre-commit = "^3.7.0"
asyncpg = "^0.29.0"
pandas = "^2.2.2"
pillow = "^10.4.0"
geojson = "^3.1.0"
scikit-learn = "^1.5.1"
numpy = "1.26.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
